+++ noodles-vcf/src/header.rs
@@ -2,14 +2,13 @@
 
 mod builder;
 pub mod file_format;
-mod number;
 pub mod parser;
 pub mod record;
 pub mod string_maps;
 
 pub use self::{
-    builder::Builder, file_format::FileFormat, number::Number, parser::ParseError, parser::Parser,
-    record::Record, string_maps::StringMaps,
+    builder::Builder, file_format::FileFormat, parser::ParseError, parser::Parser, record::Record,
+    string_maps::StringMaps,
 };
 
 use std::{hash::Hash, str::FromStr};