+++ noodles-vcf/src/record/info/field/value.rs
@@ -1,7 +1,7 @@
 use std::io;
 
 use crate::{
-    header::{record::value::map::info::Type, Number},
+    header::record::value::map::info::{Number, Type},
     variant::record::info::field::{value::Array, Value},
 };